// src/api-client/models/product-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spira API Documentation
 * Spira API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProductCategory } from './product-category';
// May contain unused imports in some cases
// @ts-ignore
import type { UnitOfMeasure } from './unit-of-measure';

/**
 * 
 * @export
 * @interface ProductResponseDto
 */
export interface ProductResponseDto {
    /**
     * Unique product ID.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'id': string;
    /**
     * Owning retailer.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'retailerId': string;
    /**
     * Barcode.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'barcode'?: string | null;
    /**
     * Product name.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'name': string;
    /**
     * Manufacturer brand.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'brand'?: string | null;
    /**
     * Store department.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'defaultDepartment'?: string | null;
    /**
     * Department the product belongs to.
     * @type {ProductCategory}
     * @memberof ProductResponseDto
     */
    'category': ProductCategory;
    /**
     * Product photo URL.
     * @type {string}
     * @memberof ProductResponseDto
     */
    'imageUrl'?: string | null;
    /**
     * Unit this product is normally counted in.
     * @type {UnitOfMeasure}
     * @memberof ProductResponseDto
     */
    'defaultUnit'?: UnitOfMeasure | null;
    /**
     * Average weight of one unit.
     * @type {number}
     * @memberof ProductResponseDto
     */
    'averageUnitWeightKg'?: number | null;
    /**
     * When the product was created (ISO 8601).
     * @type {string}
     * @memberof ProductResponseDto
     */
    'createdAt': string;
    /**
     * When the product was last modified (ISO 8601).
     * @type {string}
     * @memberof ProductResponseDto
     */
    'updatedAt': string;
}